pyproject.toml
[tool.poetry]
name = "tg-fastapi-bot"
version = "0.1.0"
description = "Dating Telegram bot using Aiogram and FastAPI"
authors = ["Tyapkova Viktoria <tyapkova06@bk.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.4"
fastapi = "^0.110"
uvicorn = "^0.29"
python-dotenv = "^1.0"
sqlalchemy = "^2.0"
asyncpg = "^0.29"
aio-pika = "^9.5.5"
redis = "^5.0.1"
pydantic-settings = "^2.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
